Text changes and bug fixed
Had some problems with setting up my domain but i've managed to set it up. Website is now live at nebicanlioglu.nl & nebicanlioglu.com! I've fixed some bug in the tech (balls) section where a language isn't rendering, this has been fixed, also the text for the skills tab has been changed to something more readable and not to very long. For now, this will be used as my main portfolio page and ofcourse will be updated in the future.